==> src/01_top3_youtube_video_crawler.ipynb <==
from dotenv import load_dotenv
load_dotenv()

from ids import extract_video_id

urls = [
    # 이세계 아이돌 
    "https://www.youtube.com/watch?v=k5O1F_IS0mU",
    "https://www.youtube.com/watch?v=rDFUl2mHIW4",
    "https://www.youtube.com/watch?v=fff8P0kYexQ",

    # 플레이브
    "https://www.youtube.com/watch?v=T_gJpnjJsnk",
    "https://www.youtube.com/watch?si=x1kGDm5IWkb8dLXU&v=EYG4ROejmyI&feature=youtu.be",
    "youtube.com/watch?si=y1keUKV5tnDV5V99&v=8L001YT4k6k&feature=youtu.be",

    # 아이브
    "youtube.com/watch?si=y1keUKV5tnDV5V99&v=8L001YT4k6k&feature=youtu.be",
    "https://www.youtube.com/watch?si=1nYu7Hlvgl8eRGgQ&v=6ZUIwj3FgUY&feature=youtu.be",
    "https://www.youtube.com/watch?si=pA-RrdGj--kfVG_A&v=B1ShLiq3EVc&feature=youtu.be",

    # RIIZE
    "youtube.com/watch?si=qeXcW7SgUNtbaIP1&v=3_b5oKGndn0&feature=youtu.be",
    "https://www.youtube.com/watch?si=jnfaeWEXYL_ZWpyW&v=iUw3LPM7OBU&feature=youtu.be",
    "youtube.com/watch?si=cCUEC_2ivRFY7CrV&v=TVqEwvbZP30&feature=youtu.be"
]

[extract_video_id(u) for u in urls]



from channel import get_channel_id_from_handle, get_uploads_playlist_id, get_videos_from_uploads_playlist

cid = get_channel_id_from_handle("isegyeidol_ofcl")
pid = get_uploads_playlist_id(cid)
videos = get_videos_from_uploads_playlist(pid, "ISEGYE_IDOL", "isegyeidol_ofcl")
len(videos), videos[0]

from videos import fetch_video_statistics_batch, chunked

# 임의 video_id 2~3개로 테스트
test_ids = ["fff8P0kYexQ", "T_gJpnjJsnk"]

stats = fetch_video_statistics_batch(test_ids)
stats


from comments import fetch_all_comments

test_video_id = "fff8P0kYexQ"  # 댓글 많은 영상
comments = fetch_all_comments(test_video_id)

len(comments), comments[0]


from ids import extract_video_id

urls = [
    # 이세계 아이돌 
    "https://www.youtube.com/watch?v=k5O1F_IS0mU",
    "https://www.youtube.com/watch?v=rDFUl2mHIW4",
    "https://www.youtube.com/watch?v=fff8P0kYexQ",

    # 플레이브
    "https://www.youtube.com/watch?v=T_gJpnjJsnk",
    "https://www.youtube.com/watch?si=x1kGDm5IWkb8dLXU&v=EYG4ROejmyI&feature=youtu.be",
    "youtube.com/watch?si=y1keUKV5tnDV5V99&v=8L001YT4k6k&feature=youtu.be",

    # 아이브
    "youtube.com/watch?si=y1keUKV5tnDV5V99&v=8L001YT4k6k&feature=youtu.be",
    "https://www.youtube.com/watch?si=1nYu7Hlvgl8eRGgQ&v=6ZUIwj3FgUY&feature=youtu.be",
    "https://www.youtube.com/watch?si=pA-RrdGj--kfVG_A&v=B1ShLiq3EVc&feature=youtu.be",

    # RIIZE
    "youtube.com/watch?si=qeXcW7SgUNtbaIP1&v=3_b5oKGndn0&feature=youtu.be",
    "https://www.youtube.com/watch?si=jnfaeWEXYL_ZWpyW&v=iUw3LPM7OBU&feature=youtu.be",
    "youtube.com/watch?si=cCUEC_2ivRFY7CrV&v=TVqEwvbZP30&feature=youtu.be"
]

[extract_video_id(u) for u in urls]


from videos import fetch_video_statistics_batch, chunked

# 임의 video_id 2~3개로 테스트
test_ids = ["fff8P0kYexQ", "T_gJpnjJsnk"]

stats = fetch_video_statistics_batch(test_ids)
stats
