==> src/03_video_statistics_enrichment.ipynb <==
from dotenv import load_dotenv
load_dotenv()

import pandas as pd

# 파일 저장을 위한 라이브러리
from pathlib import Path

# video statistic(좋조회수/좋아요/쇼츠,롱폼 여부) 함수
from videos import (
    fetch_video_statistics_batch,
    chunked,
)

BASE_DIR = Path("..")        # src 기준
DATA_DIR = BASE_DIR / "data"
INPUT_PATH = DATA_DIR / "channel_all_videos.csv"

df_videos = pd.read_csv(INPUT_PATH)

df_videos.head(), df_videos.shape


video_ids = df_videos["video_id"].dropna().unique().tolist()

len(video_ids)

video_stats_map = {}

for chunk in chunked(video_ids, size=50):
    batch_stats = fetch_video_statistics_batch(chunk)
    video_stats_map.update(batch_stats)

len(video_stats_map)

stats_df = (
    pd.DataFrame
    .from_dict(video_stats_map, orient="index")
    .reset_index()
    .rename(columns={"index": "video_id"})
)

stats_df.head()

df_videos_full = df_videos.merge(
    stats_df,
    on="video_id",
    how="left"
)

df_videos_full.head(), df_videos_full.shape

# Unknown 1개 실시간 스트리밍 영상
df_videos_full["video_type"].value_counts()

BASE_DIR = Path("..")        # src 기준
DATA_DIR = BASE_DIR / "data"
OUTPUT_PATH = DATA_DIR / "channel_videos_with_stats.csv"

df_videos_full.to_csv(OUTPUT_PATH, index=False)

print(f"Saved to {OUTPUT_PATH}")

df_videos_full[df_videos_full["video_type"]=="UNKNOWN"]
